ZDM-479 Containerise monitoring stack + various automation improvements (#23)
## Main changes

New playbook to containerise the monitoring stack:
* Deploys the Prometheus Node Exporter as a container on the proxy nodes
(checking if Docker is already present and, if not, installing it - this
is to cover all edge cases). This container collects the system metrics
and makes them available to Prometheus.
* Installs Docker on the monitoring server if it is not already present.
* Cleans up any existing monitoring containers
* Configures and deploys Prometheus as a container:
* Configured to scrape proxy application metrics, Golang system metrics
and OS-level metrics from the proxy hosts (as before)
* Uses a named volume for metrics storage, to ensure they are persisted
regardless of the lifecycle of the Prometheus container
* The Prometheus alert rule file is from the CloudAlchemy role (added
its MIT licence alongside the file)
* Configures and deploys Grafana as a container, loading the three
pre-configured dashboards as before

Notes: 
* The new monitoring playbook is still called
`deploy_zdm_monitoring.yml` but is completely different.
* The old monitoring playbook and all its associated roles have all been
removed (hence the high number of deleted files in this PR).

Other Ansible changes:
* The `deploy_zdm_proxy.yml` playbook now checks if Docker is already
installed on the proxy hosts before installing it.
* The `rolling_update_zdm_proxy.yml` and `rolling_restart_zdm_proxy.yml`
playbooks pause for a configurable number of seconds (previously
hardcoded to 10).
* Extracted the Docker installation in a separate task that does the
same things as before and adds the post-installation steps to enable
Ubuntu to run Docker without `sudo`. This task is reused by all
playbooks installing Docker.
* Extracted some other common operations into tasks to avoid
duplication.
* Renamed configuration file to be more consistent with the others.

## Other improvements

* Changed the ZDM Utility to create the Ansible Control Host Docker
container with restart policy `unless-stopped`.
* Added post-installation steps to the convenience script that installs
Docker on Ubuntu (this can be used when preparing the jumphost to run
the ZDM Utility). These new steps allow the ubuntu user to run Docker
without having to use sudo (in line with the recommendations in the
documentation).
* Terraform:
  * Replaced `whitelisted` with `allowed` in variable names
  * Removed IP range restrictions that are no longer needed.