#42 Fix docker group (#43)
I've tested the change on a brand-new set of proxy hosts, as well as
3-nodes that were installed with proxy containers before, and after
using the following commands to remove the existing docker installation:
```
sudo systemctl stop docker.service docker.socket; sudo apt -y remove docker-ce docker-ce-cli containerd.io
```
In both situations, I'm able to see ubuntu user is keeping the primary
group as ubuntu, while docker is added as an additional group.

Note, I also updated the docker detection script to use docker on the
host to run a hello-world test container. The previous approach of using
"docker --version" can only prove docker-ce-cli is installed, but if
docker engine is not working, the previous approach will not be able to
detect that.