Fix rolling restart playbook (#55)
This PR fixes a recent regression that appears to have been caused by a
change in behaviour of the Ansible Docker module. This broke the
`rolling_restart_zdm_proxy.yml` playbook (all other playbooks were not
affected).

Changes in this PR are:
- The rolling restart operation is now done by running the `docker
restart` shell command, rather than using the Docker module.
- The Ansible Docker module version has been pinned to `2.7.0`, which is
the latest version available through `pip` and compatible with the
version of Ansible used by the automation.

No documentation changes are needed for this PR, as these are purely
internal fixes that do not change the user interaction.