[Terraform] Add optional Owner tag to infrastructure resources (#58)
When creating ZDM infrastructure in AWS using the Terraform automation
for multiple deployments, it may be difficult to distinguish the
different sets of infrastructure for each deployment. The custom suffix
appended to the name of the resources helped with this, but filtering
was still quite cumbersome.

To make things easier, an Owner tag is now automatically assigned to
each provisioned resource when the user specifies a custom suffix.

Here is how it works:
- The user specifies `custom_name_suffix` in the `run_terraform_zdm.sh`
script (or the alternative version without built-in peering).
- A variable called `custom_name_suffix` is passed to Terraform,
prepended with a dash. This variable will be appended to the name of
each resource.
- A variable called `owner` is passed to Terraform, set to the same
value as `custom_name_suffix` but without prepending a dash. This will
be set as an AWS tag called Owner.

The user can then see all the resources for a certain deployment by
filtering by Owner.

If `custom_name_suffix` is left empty, no suffix is appended and the
Owner tag is left empty.