ZDM Util: upgrade Go version and some of its dependencies (#70)
This PR only affects the ZDM Util.

The changes include an upgrade of the Go version from 1.16 to 1.21.7,
and upgrading several dependencies that were outdated. In particular,
the runc transitive dependency has been upgraded from 1.1.4 to 1.1.12 to
address a security vulnerability.

Additionally, there is a fix for a small compilation error in a test
which had gone unnoticed since the changes to make the ZDM Util
OS-agnostic.

The automated tests are passing and a manual smoke test was also
successful.